Fix issue where new task's delete don't work
- Newly added tasks' delete button didnt work for
since they werent included in the deleteBtn
querySelectAll, instead, event listener was
attached to container and event delegation was
used.

+++ src/init-all-page.js
@@ -1,10 +1,18 @@
 import deleteTask from "./delete-task.js";
 
 export default function initAllPage() {
-  const deleteBtn = document.querySelectorAll(".delete-btn");
-  deleteBtn.forEach((btn) =>
-    btn.addEventListener("click", function (e) {
-      deleteTask(e.target);
-    })
-  );
+
+  // Didnt work for newly added tasks, since they werent included in the deleteBtn
+  // const deleteBtn = document.querySelectorAll(".delete-btn");
+  // deleteBtn.forEach((btn) =>
+  //   btn.addEventListener("click", function (e) {
+  //     deleteTask(e.target);
+  //   })
+  // );
+
+  const main = document.querySelector("#main");
+  main.addEventListener("click", function (e) {
+    if (!e.target.classList.contains("delete-btn")) return;
+    deleteTask(e.target);
+  });
 }
